+++ example/lib/pages/list_view_page.dart
@@ -4,20 +4,16 @@ import 'package:tuple/tuple.dart';
 import 'package:fl_paging/fl_paging.dart' as paging;
 class ListViewDataSource extends paging.PageKeyedDataSource<int, Note> {
   NoteRepository noteRepository;
-  ListViewDataSource(this.noteRepository) {
-    isEndList = false;
+
+  ListViewDataSource(this.noteRepository);
+
+  @override
+  Future<Tuple2<List<Note>, int>> loadInitial(int pageSize) async {
+    return Tuple2(await noteRepository.getNotes(0), 1);
   }
 
   @override
-  Future<Tuple2<List<Note>, int>> loadInitial() async {
-    final result = await noteRepository.getNotes(0);
-    return Tuple2(result, 0);
-  }
-
-  @override
-  Future<Tuple2<List<Note>, int>> loadPageAfter(int params) async {
-    if (params == 6) isEndList = true;
-    final result = await noteRepository.getNotes(params + 1);
-    return Tuple2(result, params + 1);
+  Future<Tuple2<List<Note>, int>> loadPageAfter(int params, int pageSize) async {
+    return Tuple2(await noteRepository.getNotes(params), params + 1);
   }
 }