+++ example/lib/pages/sliver_app_bar_list_view_page.dart
@@ -42,7 +42,7 @@ class _SliverAppBarListViewPageState extends State<SliverAppBarListViewPage> {
               )
             ];
           },
-          body: paging.ListView<Note>(
+          body: paging.PagingListView<Note>(
             padding: EdgeInsets.all(16),
             itemBuilder: (context, data, child) {
               return NoteWidget(data);