+++ lib/fl_paging.dart
@@ -1,5 +1,5 @@
 library fl_paging;
 
 export 'src/datasource/page_keyed_data_source.dart';
-export 'src/widgets/grid_view.dart';
-export 'src/widgets/list_view.dart';
+export 'src/widgets/paging_grid_view.dart';
+export 'src/widgets/paging_list_view.dart';