+++ lib/src/datasource/data_source.dart
@@ -1,6 +1,9 @@
-abstract class DataSource<Value> {
+abstract class DataSource<D> {
   /// Request load page for Data Source
-  Future<List<Value>> loadPage({bool isRefresh});
+  Future<List<D>> loadPage({bool isRefresh});
 
+  /// true when your list is don't have any data for next request.
   bool get isEndList;
 }
+
+const kDefaultPageSize = 20;