+++ lib/src/datasource/index_data_source.dart
@@ -3,7 +3,7 @@ class IndexDataSource {
   final int pageSize;
   final int initIndex;
 
-  IndexDataSource({this.pageSize = 20, this.initIndex = 20});
+  IndexDataSource({this.pageSize = 20, this.initIndex = 0});
 }
 
 class LoadParamsIndex {