+++ lib/src/widgets/default/empty_widget.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/widgets.dart';
+
+class EmptyWidget extends StatelessWidget {
+  const EmptyWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text('No Data')
+    );
+  }
+}