+++ lib/src/widgets/paging_state.dart
@@ -6,6 +6,8 @@ part 'paging_state.freezed.dart';
 abstract class PagingState<D> with _$PagingState<D> {
   const factory PagingState(List<D> datas, bool isLoadMore, bool isEndList) =
       PagingStateData;
+
   const factory PagingState.loading() = PagingStateLoading;
+
   const factory PagingState.error(dynamic error) = PagingStateError;
 }